feat: change return type of nancy_index in get_label function to long for NANCY_HIGH modes

+++ ulcerative_colitis/data/datasets/labels.py
@@ -45,7 +45,7 @@ def get_label(slide_metadata: pd.Series, mode: LabelMode) -> torch.Tensor:
         case LabelMode.NANCY_LOW | LabelMode.NANCY_LOW_ALL:
             return torch.tensor(slide_metadata["nancy_index"].item()).float()
         case LabelMode.NANCY_HIGH | LabelMode.NANCY_HIGH_ALL:
-            return torch.tensor(slide_metadata["nancy_index"].item()).int()
+            return torch.tensor(slide_metadata["nancy_index"].item()).long()
         case LabelMode.ULCERATION | LabelMode.ULCERATION_ALL:
             return torch.tensor(slide_metadata["ulceration"].item()).float()
